+++ CTF/src/main/java/ca/mcgill/science/ctf/fragments/base/BasePrintJobFragment.java
@@ -5,6 +5,7 @@ import android.support.annotation.NonNull;
 import java.util.ArrayList;
 import java.util.List;
 
+import ca.allanwang.capsule.library.swiperecyclerview.adapters.AnimationAdapter;
 import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ISwipeRecycler;
 import ca.mcgill.science.ctf.api.PrintData;
 import ca.mcgill.science.ctf.iitems.PrintJobItem;
@@ -15,12 +16,17 @@ import ca.mcgill.science.ctf.iitems.PrintJobItem;
 public abstract class BasePrintJobFragment extends BaseFragment<PrintJobItem, List<PrintData>> {
 
     @Override
+    protected void configAdapter(AnimationAdapter<PrintJobItem> adapter) {
+        super.configAdapter(adapter);
+        adapter.withItemEvent(new PrintJobItem.PrintJobClickEvent());
+    }
+
+    @Override
     protected void onResponseReceived(@NonNull List<PrintData> body, ISwipeRecycler.OnRefreshStatus onRefreshStatus) {
         List<PrintJobItem> items = new ArrayList<>();
         for (PrintData print : body)
             items.add(new PrintJobItem(print));
         mAdapter.add(items);
-        mAdapter.withItemEvent(new PrintJobItem.PrintJobClickEvent());
     }
 
     @Override
@@ -29,7 +35,6 @@ public abstract class BasePrintJobFragment extends BaseFragment<PrintJobItem, Li
         for (PrintData print : body)
             items.add(new PrintJobItem(print));
         mAdapter.setNewList(items);
-        mAdapter.withItemEvent(new PrintJobItem.PrintJobClickEvent());
     }
 
 }