+++ CTF/src/main/java/ca/mcgill/science/ctf/utils/Utils.java
@@ -17,4 +17,8 @@ public class Utils {
         NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
         return activeNetworkInfo != null && activeNetworkInfo.isConnectedOrConnecting();
     }
+
+    public static String stringFormatter(Context context, @StringRes int base, boolean condition, @StringRes int yes, @StringRes int no) {
+        return String.format(context.getString(base, condition ? context.getString(yes) : context.getString(no)));
+    }
 }